test-requirements.txt

# -- freeze time
freezegun==0.3.11

# -- allows specifying python project path to be discovered
pytest-pythonpath==0.7.1

# -- provides injectable mocker fixture
pytest-mock==0.10.1

# -- human friendly testing of http requests to external APIs
# DOCS: https://httpretty.readthedocs.io/en/latest/
httpretty==0.9.5

# -- fake data
fake-factory==0.5.6
